Move examples into example directory
