+++ examples/exampleDecorator.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const hmac = require('.')
+const hmac = require('../')
 
 module.exports = function (fastify, options, next) {
   fastify.register(hmac, {
@@ -8,7 +8,8 @@ module.exports = function (fastify, options, next) {
     getAlgorithm: () => 'sha512',
     getSignatureEncoding: () => 'base64'
   })
-  fastify.addHook('preValidation', (request, reply, next) => {
+
+  fastify.decorate('verifyHMAC', function (request, reply, next) {
     try {
       request.validateHMAC(request, reply, next)
     } catch (err) {
@@ -18,7 +19,16 @@ module.exports = function (fastify, options, next) {
 
   fastify.post('/', (req, reply) => {
     reply.type('application/json')
-    reply.send({ hello: 'hmac' })
+    reply.send({ hello: 'no validation needed' })
   })
+
+  fastify.post('/foo',
+    {
+      preValidation: [fastify.verifyHMAC]
+    },
+    (req, reply) => {
+      reply.type('application/json')
+      reply.send({ hello: 'validated HMAC' })
+    })
   next()
 }
